+++ CHANGELOG.md
@@ -6,10 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres
 to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [3.0.1][3.0.1]
+## [3.1.0][3.1.0]
+
+This release introduces Unzer PayPal Express in Java SDK.
 
 ### Added
 
+* Added Paypal-Express support:
+  * Extended `AdditionalTransactionData` with `paypal.checkoutType` field. See `com.unzer.payment.models.AdditionalTransactionData.setPaypal`.
+  * Defined `RESUMED` value for `com.unzer.payment.AbstractTransaction.Status`
+  * Added `com.unzer.payment.Unzer::updateCharge()`, `com.unzer.payment.Unzer::updateAuthorization()` which must be invoked after Paypal-Express transaction is resumed.
 * Added fields `orderId` and `invoiceId` to `Authorize`, `Charge` and `Cancel` transactions.
 * Added authorize support for Paypage. Use `Paypage::setAction(Paypage.Action.AUTHORIZE)`
 
@@ -280,6 +286,8 @@ This release brings Unzer Paylater Invoice payment type support to Java SDK.
     *   cancelAuthorization
 *   Remove deprecated classes
     *   RestCommunication
+
+[3.1.0]: http://github.com/unzerdev/java-sdk/compare/3.0.0..3.1.0
 
 [3.0.0]: http://github.com/unzerdev/java-sdk/compare/1.3.0.0..3.0.0